ref(back): refactor tests to use and organize user data generation with factories
